+++ Documentation/CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG
+
 ### 2026.0323.02
 
 ### New Features
-- Timestamps now available with color options
+- Timestamps now available with color options.
+- Added a History tab to prepare for a future loot browsing window.
+- Added profiles.
 
 ### Changes
 - Color option for the count and increment texts.
@@ -13,5 +16,3 @@
 
 #### Changes
 - Initial release
-
----